+++ src/context/weatherContext.js
@@ -1,15 +1,17 @@
 import { createContext, useState } from "react";
 
 export const WeatherContext = createContext();
-
+const limitFetchWeatherDataLocalStorage = JSON.parse(
+  localStorage.getItem("limitWeatherFetches" || [])
+);
 export const WeatherContextProvider = ({ children }) => {
-  const limitFetchWeatherDataLocalStorage = JSON.parse(
-    localStorage.getItem("limitWeatherFetches" || [])
-  );
   const [cityInput, setCityInput] = useState("");
   const [error, setError] = useState(false);
   const [fetchesLimitError, setFetchesLimitError] = useState(false);
   const [trackUserFetches, setTrackUserFetches] = useState(null);
+  const [hoursVisible, setHoursVisible] = useState(false);
+  const [emptyInput, setEmptyInput] = useState(false);
+
   const [storage, setStorage] = useState(
     limitFetchWeatherDataLocalStorage || []
   );
@@ -27,6 +29,10 @@ export const WeatherContextProvider = ({ children }) => {
         setFetchesLimitError,
         storage,
         setStorage,
+        hoursVisible,
+        setHoursVisible,
+        emptyInput,
+        setEmptyInput,
       }}
     >
       {children}
